Address module boundary review feedback
